+++ .env.default
@@ -6,3 +6,4 @@ FREEE_CLIENT_ID=dummy
 FREEE_CLIENT_SECRET=dummy
 FREEE_ENCODE_CALLBACK_URL=dummy
 FREEE_COMPANY_ID=dummy
+FREEE_AUHTORIZATION_CODE=dummy
