+++ README.md
@@ -1,3 +1,35 @@
 # attendance-at-freee-using-slack
 
 Slack スタンプを利用した freee での出席
+
+## 開発
+
+### 依存
+
+- node.js=20.9.0
+
+### 初回
+
+```sh
+copy .env.default .env #.envに変数を渡す
+yarn install
+```
+
+### local実行
+
+```sh
+yarn dev
+```
+
+### lint
+
+```sh
+yarn fmt
+```
+
+### production build
+
+```sh
+yarn build
+yarn start
+```
