+++ src/utill/freerAuthorize.ts
@@ -2,7 +2,7 @@ const token_url = 'https://accounts.secure.freee.co.jp/public_api/token'
 const redirect_uri = process.env.FREEE_ENCODE_CALLBACK_URL!
 const client_id = process.env.FREEE_CLIENT_ID!
 const client_secret = process.env.FREEE_CLIENT_SECRET!
-let code = '取得した認可コード'
+const code = process.env.FREEE_AUHTORIZATION_CODE!
 let access_token: string | null = null
 let refresh_token: string | null = null
 let lastTokenTime: number | null = null
