+++ src/utill/slackClient.ts
@@ -5,6 +5,7 @@ async function getEmailAdressFromSlack(
   app: App<StringIndexed>,
   user: string
 ): Promise<string> {
+  //memo: userに紐づくemilaは必ず取得できるはず
   const resp = await app.client.users.info({ user: user })
   const email = resp.user?.profile?.email!
   return email
